requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.4.0=pypi_0
catkin-pkg=1.0.0=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
cmake=3.28.1=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.8=pypi_0
decorator=5.1.1=pypi_0
detr=0.0.0=dev_0
diffusers=0.30.0=pypi_0
distro=1.9.0=pypi_0
dm-control=1.0.16=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
docutils=0.20.1=pypi_0
egl-probe=1.0.2=pypi_0
einops=0.7.0=pypi_0
etils=1.3.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fasteners=0.19=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pypi_0
glfw=2.6.4=pypi_0
google-auth=2.33.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.65.4=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.24.5=pypi_0
idna=3.6=pypi_0
imageio=2.33.1=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib_resources=6.1.1=py38h06a4308_1
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.12.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
labmaze=1.0.6=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libdeflate=1.17=h5eee18b_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libwebp-base=1.3.2=h5eee18b_0
lit=17.0.6=pypi_0
lxml=5.1.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markupsafe=2.1.4=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-base=3.7.2=py38h1128e8f_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpld3=0.5.9=py38h06a4308_0
mpmath=1.3.0=pypi_0
mujoco=3.1.1=pypi_0
mujoco-py=2.1.2.14=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
numpy-base=1.24.3=py38h31eccc5_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openjpeg=2.4.0=h9ca470c_1
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py38h06a4308_0
pluggy=1.5.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.2=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py=1.11.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.17.2=pypi_0
pyopengl=3.1.7=pypi_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
pytest=8.2.1=pypi_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2024.7.24=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
retry=0.9.2=pypi_0
robomimic=0.3.0=pypi_0
rospkg=1.5.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.4=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.4.0=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=2.0.0=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.1=pypi_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.1.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
